Closes #26. Initialized the database schema. Added the scaffolding for all of the domain classes, including database migrations.
I already ran them on the server, so there is no need to run them if you are developing there. However, if you are developing locally, you will need to run them with 'rake db:create db:migrate'.

You will need to recreate the 'database.yml' file to be able to connect to the database. Use google to ensure you have the right connection information.