Closes #96. Added prompts to Levels, refactored routing.
Instructors can more easily edit and view levels using the default Level#edit and Level#show actions, and instead use Level#play to play a level.

Level prompts are to provide the instructions for a level. They may provide some backstory, or simply contain step-by-step walkthroughs. They are displayed above the code editor that students will use while playing a level.

Prompts *will* be markdown enabled, but they are not as of yet.